Move file with no changes
